Added DAO objects for study, study and sample forms tied up with front end with tests.
Setting up database and flask backend is crucial for any action or tables to be working.

A somewhat large scale rework of samples forms, added study forms, should be tieding up with front end forms now. Including CRUD methods for all forms, a relation set up for study and sample, where on create, all samples will need to be assign to a study, this means sample can not exist alone.

No side effects so far. However, for testing, it is crucial that deleting study and the rest of the study manipulation should not be done at the same time.