Added dialog selectors to add a new column to table
Previous implementation allowed for a column name to be used multiple times.
It was also unclear if the columns were being added, so changes have been made to rectify this.

An addColumn function has been change to ensure duplidated columns cannot be made.
Selectors have been added to MUI dialog - a pop-up on the page - allowing columns to be selected from a list.

No other intended side effects.
Co-authored-by: Mark Turnbull (student) <2482252T@student.gla.ac.uk>
Co-authored-by: Daniel Blain (student) <2479176B@student.gla.ac.uk>