Deleted a nested beatson-app folder
This was a mistaken folder and was not necessary, there was a mix up when
trying to copy changes from branch 52 and this was a result of that mix up.
This deletion is to prevent mix up of files and to ensure the next.js is properly configured and that changes are made
in the appropriate places going foward.