WIP: Creates files to show multiple uploads formats
This change will allow effective demostration of the two possible implementations that the sample
form could have, this will allow us to effectivley showcase to the customer what options they
have in direction of the form.

It adds an additional upload page to demonstrate an alternative form submission for samples.
The page does not work with intended funcitonality as of yet but this commit
creates the methods and comments on desired implementations
the page has also been add to the navbar.

There are no side affects. This branch is to be isolated from the rest of the code base. It itself will not be merged
and instead the desired implementation chosen by the customer will be implemented in another branch.

Co-authored-by: Daniel Blain (student) <2479176B@student.gla.ac.uk>
Co-authored-by: Alyson Dick (student) <2084268D@student.gla.ac.uk>