Changed TypeScript Version and Renamed Files and Functions
Changes the Typescript version back to es5 as the features included in es2015 are not required and this version could cause compatability errors.

Also changed name of postFormToUrlAsJson.js within services folder in client to BackendAPI.js.
Name of function postFormToUrlAsJson() within BackendAPI.js changed to PostFormAsJSON() as function name was too long.

Functions and files are set to more appropriate names and we resolve possible compatability errors with regards to the version of typescript

Removed features seen in the es2015 version of typescript