Reintroduction of logic removed by mistake and files mistakenly deleted
Included logic to add the top late or the git dir to the path varible as no way to include test modules without it has been found as of writing.
Reintroduced __init__.py to test dir as it was deleted by mistake.

Reintrodudec deleted code and file

Allows for tests to be called in the same way you would a module. Allows for main application to refernce Infratructurer as a module.