WIP: Added functionality to Checkbox onClick.
Takes steps to create working checkboxs for sample characteristic addition.
This is an alternative to the menu selection implementation in StudyForm.js.

Adds a helper method changeChecked to get closer to this overall desired functionality.

Side effects are that functionality to add columns is not yet possible.
LoginForm.js was changed due to a syntax error in styles import.

Co-authored-by: Daniel Blain (student) <2479176B@student.gla.ac.uk>
Co-authored-by: Alyson Dick (student) <2084268D@student.gla.ac.uk>