Full checkbox functionality added. Users need to be able to add and remove sample characteristics when required.
Using checkboxes that correspond to each possible characteristic that
can be used is rendered on page, when checked it will add to the sample form and if it is unchecked it will remove/not add the column.

Changes to tsconfig.json to update the version a hass been made. This could affect other parts
of the current webpages further developement, however not initially noticed.

Co-authored-by: Daniel Blain (student) <2479176B@student.gla.ac.uk>
Co-authored-by: Alyson Dick (student) <2084268D@student.gla.ac.uk>