Added experiments to study form as per Mayank's request.
This change was necessary to comply with customer's needs.

The change addresses the issue by adding an experiment form as a component alongside a unique identifier based off accession for association purposes.

Side effects are that the page visually changes the layout of the study form. Also note that all of these forms appear in sequence after submission of the prior form.

Co-authored-by: Alyson Dick (student) <2084268D@student.gla.ac.uk>
Co-authored-by: Aubrey Agub <ilmmbis@users.noreply.github.com>