WIP Attempt populating rows with additionalColumn Current development plan involves having all rows in DataGrid populated with a null value, to allow editing. This is doable on the uploadForm implementation for the pre-defined columns, but not additional columns. WIP commited so that this can be picked up again. Commit also includes further development on the updateCell function in sampleForm.
Submit function in uploadForm has been amended to add the columnHeader
with a null value - this has been commented out.
updateCell now has a for loop which correctly matches
the row of the cell that has been clicked on - we hope that
this can be used to update the row in future.

No other files/functionality affected.

Co-authored-by: Mark Turnbull (student) <2482252T@student.gla.ac.uk>