Work in progress, tried having generic dao object work for study
Test run for generic object, this is the most important part of the refactoring, and to make all json works for backend and make session/json calls easy work manage.

A generic function, and everything else to work around it for study data access to database works

This will break almost all code.