Work in progress: Generic function for creating nodes are in place and working for all types. Attemp on fixing duplicate nodes also
Test run for generic object, this is the most important part of the refactoring, and to make all json works for backend and make session/json calls easy work manage.

A generic function for creating nodes, and everything else to work around it for all kinds of type data access to database works

This may break some of the code. However, so far broke codes I know has been removed.