[WIP] Updating Pop Script to Add Experiments and Samples
This is needed as only studies are being created in the current script.
This was done by factoring out the population section of the code and reusing it for experiments and samples.
Side effects - separate pop scripts for exp and samples are no longer required - everyhting is done in pop script. WIP as experiments as samples nodes are not being generated - need to be debugged.