Updated .gitignore to inlude eslintrc
This file was ment to be added to the gitignore as it was secret (.filename) and was realted to the linter, which could has resulted on linting errors appering on some but not all clients in this case.

.eslintrc was added to gitignore

This means that the linter loses the ability to be synchronised via git, could lead to people using diffrent linter configs which could effect style at worst case.