Added cell validation
Added a function that lists the cells that have undefined values and notes them as errors when the form is attempting to be submitted

this function takes the rows, itterates though them and looks at each key-pair value, notes row, column (key) and notes its as undefined in an object.
this object can then be handle with in the form of a pop up or a log viewport on the page.

button to log row state added to samples, this is used for testing and will be removed before merge

not greater changes to code base that could effect other areas outside of SampleForms.js