Sample form upload feature working.
Allows users to submit their sample data they find in their experiments.
Using similar methods to Study and Form submission this feature makes use of sendJsonToFlask.
It iterativly loops through rows and creates valid json to send to flask.
This is then populated to neo4j using the /create_sample route.

There has to be changes made to the column field names in the columns objects. No side effect are noticed yet however,this should be noted for future dev.