Added datasetDao
Added datasetDao which contains an '__init__' function and 'get_sample' function which takes dataset ID and returns the sample assosiated with that dataset, this assumes that datasets and samples are 1-1 and could be corrected later

the init function assigns itself with a driver so it can run db sessions. the get_sample function takes dataset ID and returns all the samples that are assosiated with that dataset (assuming that would only be one dataset rn), it then serializes the returned records and retuns results

no greater effect atm, no code to call these functions in place.