[WIP] Enabled Dynamic Routing + Developing Study Page
Necessary to enable clicking of study row, and to display the page for the study itself.
Used Router to send data from Study Discovery Page to (dedicated) Study page. Use fetch functionality to get experiment and sample data from database [wip].
Side effects - Now able to click a row to render page for it. Function for displaying experiment and sample data still to be fixed.

Co-authored-by: Alyson Dick (student) <2084268D@student.gla.ac.uk>
Co-authored-by: Mark Turnbull (student) <2482252T@student.gla.ac.uk>