Added working unit tests for data sets and other fixes
Added a file containing 3 unit test and a method in the allDeleteTests.py along with fixes to datasets to ammend errors and a minor fix to genericDao to ammend spelling mistake

These three unit tests test creation, updates and retreval of datasets and the deletion of datasets is handled in allDeleteTests.py. Ammendments made to datasets route file to fix issues resulting in test failures. small spelling mistake resulting in the update_dataset test to fail.

Datasets can now be added to the pipeline.