Testing SampleForm and making sure eveything passes
Th commit solves test for the Datagrid, Checkboxes, also when adding new samples, and a Submit button.

So this code was added in a file __tests__ and created SampleForm.test.js to test the SampleForm. In the code I was making sure that all of the passes and works fine, For example the DataGrid with their static columns I made sure that it should alwys be there.

And based on the side effects. There is no side effects in the code