Testing the ExperimentForm
The commit solves for testing the Expermiment Form. The Title, Description, Number of Samples and update them.

The code was about testing the ExperimentForm and making sure that eveything is fine, for example the Title, Description, Number of Samples, and also a button that going to Create Sample Form and updates the test field. And also sendJsonToFlask with the corret data. This code was add in the file test and it should pass when running npm test.

No side effects on any of the other files or functionality.