remove experiment and sample submit button onclick needed to ensure users do not create more than one iteration of the form they submit ie they only submit one study form, the exact amount of experiments and samples described and do not produce doubles when submitting a form.
adding to experiment and sample forms the correct rendering dunctionality to ensure when node
is generated that the button used for submitting the respective form is removed.
Also made the layout of StudyForm component more in line with our idea for its refactor - larger description text field and correct spacing.
No other functionality affected.